requirements.txt
requests 
bs4 
pandas 
lxml 
html5lib
sqlalchemy
psycopg2-binary
